cas: smb1398: Fix pointer-to-int-cast warning in smb1398_probe
../drivers/power/supply/cas_qcom/smb1398-charger.c: In function 'smb1398_probe':
../drivers/power/supply/cas_qcom/smb1398-charger.c:3754:30: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
 3754 |         chip->div2_cp_role = (int)of_device_get_match_data(chip->dev);
      |                              ^

+++ drivers/power/supply/cas_qcom/smb1398-charger.c
@@ -3751,7 +3751,7 @@ static int smb1398_probe(struct platform_device *pdev)
 	platform_set_drvdata(pdev, chip);
 
 	chip->standalone_mode = true;
-	chip->div2_cp_role = (int)of_device_get_match_data(chip->dev);
+	chip->div2_cp_role = (uintptr_t)of_device_get_match_data(chip->dev);
 	switch (chip->div2_cp_role) {
 	case DIV2_CP_MASTER:
 		rc = smb1398_div2_cp_master_probe(chip);
